perf(list-pages): make /blog and /work statically prerendered
Move all searchParams-driven state (search, filters, sort, pagination,
modal) into new client components so the RSC pages have no dynamic
dependencies. Next.js now reports both routes as ○ (Static).

Also moves sortByOrder from lib/content.ts into lib/sorting.ts so that
sorting utilities are safe to bundle in client components without pulling
in the fs-dependent content layer.
