+++ lib/sorting.ts
@@ -1,4 +1,12 @@
-import { sortByOrder } from "@/lib/content";
+export function sortByOrder<T extends { order?: number }>(items: T[]): T[] {
+  return [...items].sort((a, b) => {
+    if (a.order !== undefined && b.order !== undefined)
+      return a.order - b.order;
+    if (a.order !== undefined) return -1;
+    if (b.order !== undefined) return 1;
+    return 0;
+  });
+}
 
 export interface SortOption {
   label: string;
